use new layout for share document scene
