+++ app/routes/authenticated.tsx
@@ -9,8 +9,8 @@ import Home from "~/scenes/Home";
 import Search from "~/scenes/Search";
 import Templates from "~/scenes/Templates";
 import Trash from "~/scenes/Trash";
+import Layout from "~/components/AuthenticatedLayout";
 import CenteredContent from "~/components/CenteredContent";
-import Layout from "~/components/Layout";
 import PlaceholderDocument from "~/components/PlaceholderDocument";
 import Route from "~/components/ProfiledRoute";
 import SocketProvider from "~/components/SocketProvider";
