+++ app/scenes/Document/Shared.tsx
@@ -1,15 +1,12 @@
 import { Location } from "history";
 import { observer } from "mobx-react";
-import { MenuIcon } from "outline-icons";
 import * as React from "react";
 import { RouteComponentProps } from "react-router-dom";
-import styled, { useTheme } from "styled-components";
-import breakpoint from "styled-components-breakpoint";
+import { useTheme } from "styled-components";
 import DocumentModel from "~/models/Document";
 import Error404 from "~/scenes/Error404";
 import ErrorOffline from "~/scenes/ErrorOffline";
-import Button from "~/components/Button";
-import Flex from "~/components/Flex";
+import Layout from "~/components/Layout";
 import Sidebar from "~/components/Sidebar/Shared";
 import useStores from "~/hooks/useStores";
 import { NavigationNode } from "~/types";
@@ -38,9 +35,8 @@ function SharedDocumentScene(props: Props) {
   const { documents } = useStores();
   const { shareId, documentSlug } = props.match.params;
 
-  const sidebarCollapsed = ui.sidebarCollapsed || !response?.sharedTree;
-
   // ensure the wider page color always matches the theme
+  // TODO: I don't know what this does
   React.useEffect(() => {
     window.document.body.style.background = theme.background;
   }, [theme]);
@@ -68,88 +64,21 @@ function SharedDocumentScene(props: Props) {
     return <Loading location={props.location} />;
   }
 
-  return (
-    <Container auto>
-      <MobileMenuButton
-        onClick={ui.toggleMobileSidebar}
-        icon={<MenuIcon />}
-        iconColor="currentColor"
-        neutral
-      />
+  const sidebar = response.sharedTree ? (
+    <Sidebar rootNode={response.sharedTree} shareId={shareId} />
+  ) : undefined;
 
-      {response.sharedTree && (
-        <Sidebar rootNode={response.sharedTree} shareId={shareId} />
-      )}
-      <Content
-        auto
-        justify="center"
-        $isResizing={ui.sidebarIsResizing}
-        $sidebarCollapsed={sidebarCollapsed}
-        style={
-          sidebarCollapsed
-            ? undefined
-            : {
-                marginLeft: `${ui.sidebarWidth}px`,
-              }
-        }
-      >
-        <Document
-          abilities={EMPTY_OBJECT}
-          document={response.document}
-          sharedTree={response.sharedTree}
-          shareId={shareId}
-          readOnly
-        />
-      </Content>
-    </Container>
+  return (
+    <Layout sidebar={sidebar}>
+      <Document
+        abilities={EMPTY_OBJECT}
+        document={response.document}
+        sharedTree={response.sharedTree}
+        shareId={shareId}
+        readOnly
+      />
+    </Layout>
   );
 }
-
-// XXX this is copy paste and should be factored out into its own thing, along with the usage in layout
-const Container = styled(Flex)`
-  background: ${(props) => props.theme.background};
-  transition: ${(props) => props.theme.backgroundTransition};
-  position: relative;
-  width: 100%;
-  min-height: 100%;
-`;
-
-const MobileMenuButton = styled(Button)`
-  position: fixed;
-  top: 12px;
-  left: 12px;
-  z-index: ${(props) => props.theme.depths.sidebar - 1};
-
-  ${breakpoint("tablet")`
-    display: none;
-  `};
-
-  @media print {
-    display: none;
-  }
-`;
-
-const Content = styled(Flex)<{
-  $isResizing?: boolean;
-  $sidebarCollapsed?: boolean;
-}>`
-  margin: 0;
-  transition: ${(props) =>
-    props.$isResizing ? "none" : `margin-left 100ms ease-out`};
-
-  @media print {
-    margin: 0 !important;
-  }
-
-  ${breakpoint("mobile", "tablet")`
-    margin-left: 0 !important;
-  `}
-
-  ${breakpoint("tablet")`
-    ${(props: any) =>
-      props.$sidebarCollapsed &&
-      `margin-left: ${props.theme.sidebarCollapsedWidth}px;`}
-  `};
-`;
 
 export default observer(SharedDocumentScene);
